+++ packages/terra-core-docs/src/terra-dev-site/test/button/CompactButton.test.jsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import Button from 'terra-button';
+
+const CompactButton = () => <Button id="compactButton" text="Compact" isCompact />;
+
+export default CompactButton;