+++ packages/terra-core-docs/src/terra-dev-site/test/cell-grid/CellGrid.test.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import CellGrid from 'terra-cell-grid';
+
+const CellGridTest = () => (
+  <CellGrid>
+    Test Content Node
+  </CellGrid>
+);
+
+export default CellGridTest;