+++ packages/terra-core-docs/src/terra-dev-site/test/demographics-banner/DeceasedDemographicsBannerNoLabel.test.jsx
@@ -1,0 +1,4 @@
+import React from 'react';
+import DemographicsBanner from 'terra-demographics-banner';
+
+export default () => <DemographicsBanner id="deceased-demographics-no-label" deceasedDate="March 12, 2017" />;