+++ packages/terra-core-docs/src/terra-dev-site/test/form-checkbox/checkbox/DisabledCheckbox.test.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import Checkbox from 'terra-form-checkbox';
+
+const checkbox = () => (
+  <div>
+    <Checkbox id="disabled" labelText="Disabled Checkbox" disabled />
+    <Checkbox id="disabledchecked" labelText="Disabled and Checked Checkbox" defaultChecked disabled />
+  </div>
+);
+
+export default checkbox;