+++ packages/terra-core-docs/src/terra-dev-site/test/form-checkbox/checkbox/PopulatedCheckbox.test.jsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import Checkbox from 'terra-form-checkbox';
+
+const checkbox = () => (
+  <Checkbox
+    id="populated"
+    labelText="Do you have any Children?"
+    labelTextAttrs={{ className: 'healtheintent-control-label-text' }}
+    name="children_present"
+    value="children_present"
+    inputAttrs={{ className: 'healtheintent-application' }}
+    isInline
+  />
+);
+
+export default checkbox;