+++ packages/terra-demographics-banner/src/terra-dev-site/test/demographics-banner/DeceasedDemographicsBannerNoLabel.test.jsx
@@ -1,4 +1,0 @@
-import React from 'react';
-import DemographicsBanner from '../../../DemographicsBanner';
-
-export default () => <DemographicsBanner id="deceased-demographics-no-label" deceasedDate="March 12, 2017" />;