+++ packages/terra-demographics-banner/src/terra-dev-site/test/demographics-banner/PostMenstrualAgeDemographicsBannerNoLabel.test.jsx
@@ -1,4 +1,0 @@
-import React from 'react';
-import DemographicsBanner from '../../../DemographicsBanner';
-
-export default () => <DemographicsBanner id="post-menstrual-no-label" postMenstrualAge="April 5, 2016" />;