+++ packages/terra-form-checkbox/src/terra-dev-site/test/form-checkbox/checkbox/HiddenLabelCheckbox.test.jsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import Checkbox from '../../../../Checkbox';
-
-const checkbox = () => (<Checkbox id="hidden" labelText="can you see me?" isLabelHidden />);
-
-export default checkbox;