+++ packages/terra-overlay/CHANGELOG.md
@@ -4,6 +4,12 @@
 
 * Changed
   * Upgrade focus-trap-react to version ^8.0.0.
+  * Updated dependencies to not be hard-coded.
+
+## 3.69.0 - (November 2, 2021)
+
+* Changed
+  * Minor dependency version bump
 
 ## 3.68.0 - (September 21, 2021)